update reliably
I didn't want to have a command in my update script that will effect all docker images but `docker-compose down --rmi all` doesn't delete the image that is built in a dockerfile (I think). Deleting the old image is necessary for nextcloud because otherwise it will not reliably download the newest version.
That's why I am really cleaning up now and reinstall everything for every update.